added wrap component for build profile form
